Implemented preemptive multitasking and server abstraction

+++ kernel/include/tasks/syscalls.h
@@ -10,4 +10,4 @@ void sys_msg_send(GeneralRegisters r, uint32_t eip, uint32_t esp);
 void sys_msg_receive(GeneralRegisters r, uint32_t eip, uint32_t esp);
 void sys_print(GeneralRegisters r, uint32_t eip, uint32_t esp);
 
-void sys_call(GeneralRegisters r, uint32_t eip, uint32_t cs, uint32_t f, uint32_t esp, uint32_t ss);
+void sys_call(GeneralRegisters r, uint32_t eip, uint16_t cs, uint32_t f, uint32_t esp, uint16_t ss);
